@@ -6,44 +6,24 @@ description = 'A tilesource for large_image to composite other tile sources'
 long_description = description + '\n\nSee the large-image package for more details.'
 
 
-def prerelease_local_scheme(version):
-    """
-    Return local scheme version unless building on master in CircleCI.
-
-    This function returns the local scheme version number
-    (e.g. 0.0.0.dev<N>+g<HASH>) unless building on CircleCI for a
-    pre-release in which case it ignores the hash and produces a
-    PEP440 compliant pre-release version number (e.g. 0.0.0.dev<N>).
-    """
-    from setuptools_scm.version import get_local_node_and_date
-
-    if os.getenv('CIRCLE_BRANCH') in ('master', ):
-        return ''
-    else:
-        return get_local_node_and_date(version)
-
-
 try:
     from setuptools_scm import get_version
 
-    version = get_version(root='../..', local_scheme=prerelease_local_scheme)
+    version = get_version(root='../..')
     limit_version = f'>={version}' if '+' not in version and not os.getenv('TOX_ENV_NAME') else ''
 except (ImportError, LookupError):
     limit_version = ''
 
 setup(
     name='large-image-source-multi',
-    use_scm_version={'root': '../..', 'local_scheme': prerelease_local_scheme,
-                     'fallback_version': '0.0.0'},
     description=description,
     long_description=long_description,
     long_description_content_type='text/x-rst',
-    license='Apache Software License 2.0',
+    license='Apache-2.0',
     author='Kitware, Inc.',
     author_email='kitware@kitware.com',
     classifiers=[
         'Development Status :: 5 - Production/Stable',
-        'License :: OSI Approved :: Apache Software License',
         'Programming Language :: Python :: 3',
         'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
@@ -66,7 +46,7 @@ setup(
     keywords='large_image, tile source',
     packages=find_packages(exclude=['test', 'test.*']),
     url='https://github.com/girder/large_image',
-    python_requires='>=3.6',
+    python_requires='>=3.8',
     entry_points={
         'large_image.source': [
             'multi = large_image_source_multi:MultiFileTileSource',
