@@ -240,7 +240,7 @@ def checkForLargeImageFiles(event):  # noqa
                 pass
     # We couldn't automatically set this as a large image
     girder.logger.info(
-        'Saved file %s cannot be automatically used as a largeImage' % str(file['_id']))
+        'Saved file %s cannot be automatically used as a largeImage', str(file['_id']))
 
 
 def removeThumbnails(event):
@@ -411,7 +411,7 @@ def metadataSearchHandler(  # noqa
         return []
     filter = {'$and': filter} if len(filter) > 1 else filter[0]
     result = {}
-    logger.debug('Metadata search uses filter: %r' % filter)
+    logger.debug('Metadata search uses filter: %r', filter)
     for model in searchModels or types:
         modelInst = ModelImporter.model(*model if isinstance(model, tuple) else [model])
         if searchModels is None:
@@ -437,7 +437,7 @@ def metadataSearchHandler(  # noqa
                 if entry is not None and offset:
                     offset -= 1
                     continue
-                elif entry is not None:
+                if entry is not None:
                     result[searchModels[model]['model']].append(resultModelInst.filter(entry, user))
                     if limit and len(result[searchModels[model]['model']]) == limit:
                         break
@@ -566,7 +566,7 @@ def yamlConfigFile(folder, name, user):
         if item:
             for file in Item().childFiles(item):
                 if file['size'] > 10 * 1024 ** 2:
-                    logger.info('Not loading %s -- too large' % file['name'])
+                    logger.info('Not loading %s -- too large', file['name'])
                     continue
                 with File().open(file) as fptr:
                     config = yaml.safe_load(fptr)
@@ -658,7 +658,7 @@ def validateBoolean(doc):
     val = doc['value']
     if str(val).lower() not in ('false', 'true', ''):
         raise ValidationException('%s must be a boolean.' % doc['key'], 'value')
-    doc['value'] = (str(val).lower() != 'false')
+    doc['value'] = str(val).lower() != 'false'
 
 
 @setting_utilities.validator({
@@ -675,7 +675,7 @@ def validateBooleanOrICCIntent(doc):
         if str(val).lower() not in ('false', 'true', ''):
             raise ValidationException(
                 '%s must be a boolean or a named intent.' % doc['key'], 'value')
-        doc['value'] = (str(val).lower() != 'false')
+        doc['value'] = str(val).lower() != 'false'
 
 
 @setting_utilities.validator({
