@@ -384,6 +384,12 @@ class AnnotationSchema:
                 'exclusiveMinimum': 0,
                 'description': 'radius used for heatmap interpretation',
             },
+            'scaleWithZoom': {
+                'type': 'boolean',
+                'description':
+                    'If true, and interpreted as a heatmap, scale the size '
+                    'of points with the zoom level of the map.',
+            },
             'colorRange': colorRangeSchema,
             'rangeValues': rangeValueSchema,
             'normalizeRange': {
@@ -836,7 +842,7 @@ class Annotation(AccessControlledModel):
         """
         annotation = super().load(id, *args, **kwargs)
         if annotation is None:
-            return
+            return None
 
         if getElements:
             # It is possible that we are trying to read the elements of an
@@ -1102,7 +1108,7 @@ class Annotation(AccessControlledModel):
                     element['id'] = str(element['id'])
                 # Handle elements with large arrays by checking that a
                 # conversion to a numpy array works
-                keys = None
+                keys = {}
                 if len(element.get('points', element.get('values', []))) > VALIDATE_ARRAY_LENGTH:
                     key = 'points' if 'points' in element else 'values'
                     try:
@@ -1224,7 +1230,7 @@ class Annotation(AccessControlledModel):
                 version = oldVersions[1]['_version']
         annotation = Annotation().getVersion(id, version, user, force=force)
         if annotation is None:
-            return
+            return None
         # If this is the most recent (active) annotation, don't do anything.
         # Otherwise, revert it.
         if not annotation.get('_active', True):
