@@ -13,7 +13,11 @@ class TileSourceAssetstoreError(TileSourceError):
     pass
 
 
-class TileSourceXYZRangeError(TileSourceError):
+class TileSourceRangeError(TileSourceError):
+    pass
+
+
+class TileSourceXYZRangeError(TileSourceRangeError):
     pass
 
 
