@@ -14,49 +14,54 @@ Please note that this webserver will not work with Classic Notebook and will
 likely lead to crashes. This is only for use in JupyterLab.
 
 """
+import ast
+import asyncio
 import importlib.util
 import json
 import os
-import re
+import threading
 import weakref
 from typing import Any, Dict, List, Optional, Tuple, Union, cast
+from urllib.parse import parse_qs, quote, urlencode, urlparse, urlunparse
 
-from large_image.exceptions import TileSourceXYZRangeError
+import numpy as np
+
+import large_image
+from large_image.exceptions import TileSourceError, TileSourceXYZRangeError
 from large_image.tilesource.utilities import JSONDict
 
 ipyleafletPresent = importlib.util.find_spec('ipyleaflet') is not None
+ipyvuePresent = importlib.util.find_spec('ipyvue') is not None
+aiohttpPresent = importlib.util.find_spec('aiohttp') is not None
 
 
 class IPyLeafletMixin:
     """Mixin class to support interactive visualization in JupyterLab.
 
-    This class implements ``_ipython_display_`` with ``ipyleaflet``
-    to display an interactive image visualizer for the tile source
-    in JupyterLab.
+    This class implements ``_ipython_display_`` with ``ipyleaflet`` to display
+    an interactive image visualizer for the tile source in JupyterLab.
 
-    Install `ipyleaflet <https://github.com/jupyter-widgets/ipyleaflet>`_
-    to interactively visualize tile sources in JupyterLab.
+    Install `ipyleaflet <https://github.com/jupyter-widgets/ipyleaflet>`_ to
+    interactively visualize tile sources in JupyterLab.
 
-    For remote JupyterHub environments, you may need to configure
-    the class variables ``JUPYTER_HOST`` or ``JUPYTER_PROXY``.
+    For remote JupyterHub environments, you may need to configure the class
+    variables ``JUPYTER_HOST`` or ``JUPYTER_PROXY``.
 
     If ``JUPYTER_PROXY`` is set, it overrides ``JUPYTER_HOST``.
 
-    Use ``JUPYTER_HOST`` to set the host name of the machine such
-    that the tile URL can be accessed at
-    ``'http://{JUPYTER_HOST}:{port}'``.
+    Use ``JUPYTER_HOST`` to set the host name of the machine such that the tile
+    URL can be accessed at ``'http://{JUPYTER_HOST}:{port}'``.
 
-    Use ``JUPYTER_PROXY`` to leverage ``jupyter-server-proxy`` to
-    proxy the tile serving port through Jupyter's authenticated web
-    interface. This is useful in Docker and cloud JupyterHub
-    environments. You can set the environment variable
-    ``LARGE_IMAGE_JUPYTER_PROXY`` to control the default value of
-    ``JUPYTER_PROXY``. If ``JUPYTER_PROXY`` is set to ``True``, the
-    default will be ``'/proxy/`` which will work for most Docker
-    Jupyter configurations. If in a cloud JupyterHub environment,
-    this will get a bit more nuanced as the
-    ``JUPYTERHUB_SERVICE_PREFIX`` may need to prefix the
-    ``'/proxy/'``.
+    Use ``JUPYTER_PROXY`` to leverage ``jupyter-server-proxy`` to proxy the
+    tile serving port through Jupyter's authenticated web interface.  This is
+    useful in Docker and cloud JupyterHub environments.  You can set the
+    environment variable ``LARGE_IMAGE_JUPYTER_PROXY`` to control the default
+    value of ``JUPYTER_PROXY``.  If ``JUPYTER_PROXY`` is set to ``True``, the
+    default will be ``'/proxy/`` which will work for most Docker Jupyter
+    configurations. If in a cloud JupyterHub environment, this will get a bit
+    more nuanced as the ``JUPYTERHUB_SERVICE_PREFIX`` may need to prefix the
+    ``'/proxy/'``.  If set to ``'auto'``, an automatic value will be chosen
+    through some effort to detect the environment.
 
     To programmatically set these values:
 
@@ -80,19 +85,23 @@ class IPyLeafletMixin:
     """
 
     JUPYTER_HOST = '127.0.0.1'
-    JUPYTER_PROXY = os.environ.get('LARGE_IMAGE_JUPYTER_PROXY', False)
+    JUPYTER_PROXY: Union[str, bool] = os.environ.get('LARGE_IMAGE_JUPYTER_PROXY', 'auto')
 
     _jupyter_server_manager: Any
 
     def __init__(self, *args, **kwargs) -> None:
         self._jupyter_server_manager = None
-        self._map = Map()
+        self._map = Map(ts=self)
         if ipyleafletPresent:
             self.to_map = self._map.to_map
             self.from_map = self._map.from_map
 
     def as_leaflet_layer(self, **kwargs) -> Any:
         # NOTE: `as_leaflet_layer` is supported by ipyleaflet.Map.add
+
+        if not getattr(self, '_noCache', False):
+            msg = 'Cannot set the style of a cached source'
+            raise TileSourceError(msg)
 
         if self._jupyter_server_manager is None:
             # Must relaunch to ensure style updates work
@@ -103,6 +112,10 @@ class IPyLeafletMixin:
 
         port = self._jupyter_server_manager.port
 
+        if self.JUPYTER_PROXY == 'auto':
+            self._autoJupyterProxy()
+        if self.JUPYTER_PROXY and str(self.JUPYTER_PROXY).lower() in {'true', 'false'}:
+            self.JUPYTER_PROXY = str(self.JUPYTER_PROXY).lower() == 'true'
         if self.JUPYTER_PROXY:
             if isinstance(self.JUPYTER_PROXY, str):
                 base_url = f'{self.JUPYTER_PROXY.rstrip("/")}/{port}'
@@ -112,10 +125,19 @@ class IPyLeafletMixin:
             base_url = f'http://{self.JUPYTER_HOST}:{port}'
 
         # Use repr in URL params to prevent caching across sources/styles
-        endpoint = f'tile?z={{z}}&x={{x}}&y={{y}}&encoding=png&repr={self.__repr__()}'
+        endpoint = f'tile?z={{z}}&x={{x}}&y={{y}}&encoding=png&repr={self!r}'
         return self._map.make_layer(
             self.metadata,  # type: ignore[attr-defined]
             f'{base_url}/{endpoint}')
+
+    def _autoJupyterProxy(self):
+        if importlib.util.find_spec('google') and importlib.util.find_spec('google.colab'):
+            # colab intercepts localhost
+            self.JUPYTER_PROXY = 'https://localhost'
+        elif importlib.util.find_spec('jupyter_server_proxy'):
+            self.JUPYTER_PROXY = True
+        else:
+            self.JUPYTER_PROXY = False
 
     # Only make _ipython_display_ available if ipyleaflet is installed
     if ipyleafletPresent:
@@ -163,6 +185,7 @@ class Map:
             on the girder client.
         """
         self._layer = self._map = self._metadata = self._frame_slider = None
+        self._frame_histograms: Optional[Dict[int, Any]] = None
         self._ts = ts
         if (not url or not metadata) and gc and (id or resource):
             fileId = None
@@ -260,7 +283,7 @@ class Map:
         ipyleaflet layer, and the center of the tile source.
         """
         from ipyleaflet import Map, basemaps, projections
-        from ipywidgets import IntSlider, VBox
+        from ipywidgets import VBox
 
         try:
             default_zoom = metadata['levels'] - metadata['sourceLevels']
@@ -276,21 +299,11 @@ class Map:
             crs = dict(
                 name='PixelSpace',
                 custom=True,
-                # Why does this need to be 256?
                 resolutions=[2 ** (metadata['levels'] - 1 - l) for l in range(20)],
 
-                # This works but has x and y reversed
                 proj4def='+proj=longlat +axis=esu',
-                bounds=[[0, 0], [metadata['sizeY'], metadata['sizeX']]],
-                # Why is origin X, Y but bounds Y, X?
+                bounds=[[0, 0], [metadata['sizeX'], metadata['sizeY']]],
                 origin=[0, metadata['sizeY']],
-
-                # This almost works to fix the x, y reversal, but
-                # - bounds are weird and other issues occur
-                # proj4def='+proj=longlat +axis=seu',
-                # bounds=[[-metadata['sizeX'],-metadata['sizeY']],
-                #         [metadata['sizeX'],metadata['sizeY']]],
-                # origin=[0,0],
             )
         layer = layer or self._layer
 
@@ -311,16 +324,14 @@ class Map:
 
         children: List[Any] = []
         frames = metadata.get('frames')
-        if frames is not None:
-            self._frame_slider = IntSlider(
-                value=0,
-                min=0,
-                max=len(frames) - 1,
-                description='Frame:',
-            )
-            if self._frame_slider:
-                self._frame_slider.observe(self.update_frame, names='value')
-                children.append(self._frame_slider)
+        if frames is not None and ipyvuePresent and aiohttpPresent:
+            from large_image.widgets.components import FrameSelector
+
+            self.frame_selector = FrameSelector()
+            self.frame_selector.imageMetadata = metadata
+            self.frame_selector.updateFrameCallback = self.update_frame
+            self.frame_selector.getFrameHistogram = self.get_frame_histogram
+            children.append(self.frame_selector)
 
         m = Map(
             crs=crs,
@@ -333,12 +344,103 @@ class Map:
             dragging=True,
             attribution_control=False,
         )
+        if not self._geospatial:
+            m.fit_bounds(bounds=[[0, 0], [metadata['sizeY'], metadata['sizeX']]])
         if self._geospatial:
             m.add_layer(layer)
         self._map = m
         children.append(m)
 
+        self.add_region_indicator()
         return VBox(children)
+
+    def add_region_indicator(self):
+        from ipyleaflet import FullScreenControl, GeomanDrawControl, Popup
+        from ipywidgets import HTML
+
+        metadata = self._metadata
+        if self._map is None or metadata is None:
+            return
+
+        self.info_label = HTML()
+        popup = Popup(child=self.info_label)
+        popup.close_popup()
+        draw_control = GeomanDrawControl(
+            rectangle=dict(shapeOptions=dict(color='#19a7ff')),
+            # enable drawing other shapes by specifying styling
+            polygon=dict(),
+            circle=dict(),
+            polyline=dict(),
+            circlemarker=dict(),
+            rotate=False,
+            cut=False,
+            edit=False,
+        )
+
+        transformer = None
+        if metadata.get('geospatial') and metadata.get('projection'):
+            import pyproj
+
+            transformer = pyproj.Transformer.from_crs(
+                'EPSG:4326',
+                metadata['projection'],
+                always_xy=True,
+            )
+
+        def handle_interaction(**kwargs):
+            coords = kwargs.get('coordinates')
+            info = []
+            if self._map is None or metadata is None or coords is None:
+                return
+            x, y = [coords[1], coords[0]]
+            interaction_type = kwargs.get('type')
+            if interaction_type == 'click':
+                for rectangle in draw_control.data:
+                    rect_coords = rectangle.get('geometry', {}).get('coordinates', [[]])[0]
+                    xmin, xmax = rect_coords[0][0], rect_coords[2][0]
+                    ymin, ymax = rect_coords[0][1], rect_coords[1][1]
+                    width, height = metadata['sizeX'], metadata['sizeY']
+                    if (not (x >= xmin and x <= xmax and y >= ymin and y <= ymax)):
+                        continue
+                    if transformer is not None:
+                        roi = f'[{xmin:.8g}, {ymin:.8g}, {(xmax - xmin):.8g}, {(ymax - ymin):.8g}]'
+                        info.append(f'Lon Range: [{xmin:.8g}, {xmax:.8g}]')
+                        info.append(f'Lat Range: [{ymin:.8g}, {ymax:.8g}]')
+                        info.append(f'Lon/Lat ROI: {roi}')
+                        x0, y0 = transformer.transform(xmin, ymin)
+                        x1, y1 = transformer.transform(xmax, ymax)
+                        bounds = metadata['bounds']
+                        x0, x1 = (
+                            round((v - bounds['xmin']) /
+                                  (bounds['xmax'] - bounds['xmin']) *
+                                  width) for v in [x0, x1]
+                        )
+                        y0, y1 = (
+                            round((v - bounds['ymax']) /
+                                  (bounds['ymin'] - bounds['ymax']) *
+                                  height) for v in [y1, y0]
+                        )
+                        info.append(f'X/Y ROI: [{x0}, {y0}, {x1 - x0}, {y1 - y0}]')
+                    else:
+                        xmin, xmax = round(xmin), round(xmax)
+                        ymin, ymax = round(height - ymax), round(height - ymin)
+                        roi = f'[{xmin:.8g}, {ymin:.8g}, {(xmax - xmin):.8g}, {(ymax - ymin):.8g}]'
+                        info.append(f'X Range: [{xmin:.8g}, {xmax:.8g}]')
+                        info.append(f'Y Range: [{ymin:.8g}, {ymax:.8g}]')
+                        info.append(f'X/Y ROI: {roi}')
+                    self.info_label.value = ''.join(f'<div>{i}</div>' for i in info)
+                    popup.open_popup((rect_coords[1][1], (xmax - xmin) / 2 + xmin))
+                    if popup not in self._map.layers:
+                        self._map.add(popup)
+
+        def handle_draw(target, action, geo_json):
+            if action == 'deleted':
+                popup.close_popup()
+
+        draw_control.on_draw(handle_draw)
+        self._map.on_interaction(handle_interaction)
+        self._map.add(draw_control)
+        self._map.add(FullScreenControl())
 
     @property
     def layer(self) -> Any:
@@ -393,18 +495,69 @@ class Map:
             return transf.transform(x, y)
         return x, self._metadata['sizeY'] - y
 
-    def update_frame(self, event, **kwargs):
-        frame = int(event.get('new'))
+    def update_frame(self, frame, style, **kwargs):
         if self._layer:
-            if 'frame=' in self._layer.url:
-                self._layer.url = re.sub(r'frame=(\d+)', f'frame={frame}', self._layer.url)
-            else:
-                if '?' in self._layer.url:
-                    self._layer.url = self._layer.url.replace('?', f'?frame={frame}&')
-                else:
-                    self._layer.url += f'?frame={frame}'
-
+            parsed_url = urlparse(self._layer.url)
+            query = parsed_url.query
+            query = {k: v[0] for k, v in parse_qs(query).items()}
+            query.update(dict(
+                frame=frame,
+                style=json.dumps(style),
+            ))
+            query_string = urlencode(query, quote_via=quote, safe='{}')
+            self._layer.url = urlunparse((
+                parsed_url.scheme, parsed_url.netloc,
+                parsed_url.path, parsed_url.params,
+                query_string, parsed_url.fragment,
+            ))
             self._layer.redraw()
+
+    def get_frame_histogram(self, query):
+        import aiohttp
+
+        if self._layer is not None:
+            if self._frame_histograms is None:
+                self._frame_histograms = {}
+
+            frame = query.get('frame')
+            parsed_url = urlparse(self._layer.url)
+            query_string = urlencode(query)
+            scheme = parsed_url.scheme or 'http'
+            netloc = parsed_url.netloc
+            if not netloc and self._ts is not None:
+                netloc = f'{self._ts.JUPYTER_HOST}:{self._ts._jupyter_server_manager.port}'
+            histogram_url = urlunparse((
+                scheme, netloc,
+                '/histogram', parsed_url.params,
+                query_string, parsed_url.fragment,
+            ))
+
+            async def fetch(url):
+                async with aiohttp.ClientSession(
+                    timeout=aiohttp.ClientTimeout(total=900),
+                ) as session:
+                    async with session.get(url) as response:
+                        self._frame_histograms[frame] = await response.json()  # type: ignore
+                        # rewrite whole object for watcher
+                        self.frame_selector.frameHistograms = (
+                            self._frame_histograms.copy()  # type: ignore
+                        )
+
+            asyncio.ensure_future(fetch(histogram_url))
+
+
+# used for encoding histogram data
+class NumpyEncoder(json.JSONEncoder):
+    """Special json encoder for numpy types from https://stackoverflow.com/a/49677241"""
+
+    def default(self, obj):
+        if isinstance(obj, np.integer):
+            return int(obj)
+        if isinstance(obj, np.floating):
+            return float(obj)
+        if isinstance(obj, np.ndarray):
+            return obj.tolist()
+        return json.JSONEncoder.default(self, obj)
 
 
 def launch_tile_server(tile_source: IPyLeafletMixin, port: int = 0) -> Any:
@@ -443,6 +596,26 @@ def launch_tile_server(tile_source: IPyLeafletMixin, port: int = 0) -> Any:
             self.write(json.dumps(manager.tile_source.metadata))  # type: ignore[attr-defined]
             self.set_header('Content-Type', 'application/json')
 
+    class TileSourceHistogramHandler(tornado.web.RequestHandler):
+        """REST endpoint to get image metadata."""
+
+        async def get(self) -> None:
+            kwargs = {k: ast.literal_eval(self.get_argument(k)) for k in self.request.arguments}
+
+            def fetch():
+                if not hasattr(manager, '_histogram_semaphore'):
+                    manager._histogram_semaphore = threading.Semaphore(  # type: ignore
+                        min(6, large_image.config.cpu_count()),
+                    )
+                with manager._histogram_semaphore:  # type: ignore[attr-defined]
+                    histogram = manager.tile_source._unstyled.histogram(  # type: ignore
+                        **kwargs,
+                    ).get('histogram', [{}])
+                self.write(json.dumps(histogram, cls=NumpyEncoder))
+                self.set_header('Content-Type', 'application/json')
+
+            await tornado.ioloop.IOLoop.current().run_in_executor(None, fetch)
+
     class TileSourceTileHandler(tornado.web.RequestHandler):
         """REST endpoint to serve tiles from image in slippy maps standard."""
 
@@ -451,6 +624,9 @@ def launch_tile_server(tile_source: IPyLeafletMixin, port: int = 0) -> Any:
             y = int(self.get_argument('y'))
             z = int(self.get_argument('z'))
             frame = int(self.get_argument('frame', default='0'))
+            style = self.get_argument('style', default=None)
+            if style:
+                manager.tile_source.style = json.loads(style)  # type: ignore[attr-defined]
             encoding = self.get_argument('encoding', 'PNG')
             try:
                 tile_binary = manager.tile_source.getTile(  # type: ignore[attr-defined]
@@ -465,6 +641,7 @@ def launch_tile_server(tile_source: IPyLeafletMixin, port: int = 0) -> Any:
 
     app = tornado.web.Application([
         (r'/metadata', TileSourceMetadataHandler),
+        (r'/histogram', TileSourceHistogramHandler),
         (r'/tile', TileSourceTileHandler),
     ])
     sockets = tornado.netutil.bind_sockets(port, '')
