@@ -45,58 +45,56 @@ def numpyResize(
 ) -> np.ndarray:
     if resample_method == ResampleMethod.NP_NEAREST:
         return tile[::2, ::2]
-    else:
-        if tile.shape[0] % 2 != 0:
-            tile = np.append(tile, np.expand_dims(tile[-1], axis=0), axis=0)
-        if tile.shape[1] % 2 != 0:
-            tile = np.append(tile, np.expand_dims(tile[:, -1], axis=1), axis=1)
+    if tile.shape[0] % 2 != 0:
+        tile = np.append(tile, np.expand_dims(tile[-1], axis=0), axis=0)
+    if tile.shape[1] % 2 != 0:
+        tile = np.append(tile, np.expand_dims(tile[:, -1], axis=1), axis=1)
 
-        pixel_selection = None
-        subarrays = np.asarray(
-            [
-                tile[0::2, 0::2],
-                tile[1::2, 0::2],
-                tile[0::2, 1::2],
-                tile[1::2, 1::2],
-            ],
+    pixel_selection = None
+    subarrays = np.asarray(
+        [
+            tile[0::2, 0::2],
+            tile[1::2, 0::2],
+            tile[0::2, 1::2],
+            tile[1::2, 1::2],
+        ],
+    )
+
+    if resample_method == ResampleMethod.NP_MEAN:
+        return np.mean(subarrays, axis=0).astype(tile.dtype)
+    if resample_method == ResampleMethod.NP_MEDIAN:
+        return np.median(subarrays, axis=0).astype(tile.dtype)
+    if resample_method == ResampleMethod.NP_MAX:
+        return np.max(subarrays, axis=0).astype(tile.dtype)
+    if resample_method == ResampleMethod.NP_MIN:
+        return np.min(subarrays, axis=0).astype(tile.dtype)
+    if resample_method == ResampleMethod.NP_MAX_COLOR:
+        summed = np.sum(subarrays, axis=3)
+        pixel_selection = np.argmax(summed, axis=0)
+    elif resample_method == ResampleMethod.NP_MIN_COLOR:
+        summed = np.sum(subarrays, axis=3)
+        pixel_selection = np.argmin(summed, axis=0)
+    elif resample_method == ResampleMethod.NP_MODE:
+        # if a pixel occurs twice in a set of four, it is a mode
+        # if no mode, default to pixel 0. check for minimal matches 1=2, 1=3, 2=3
+        pixel_selection = np.where(
+            (
+                (subarrays[1] == subarrays[2]).all(axis=2) |
+                (subarrays[1] == subarrays[3]).all(axis=2)
+            ),
+            1, np.where(
+                (subarrays[2] == subarrays[3]).all(axis=2),
+                2, 0,
+            ),
         )
 
-        if resample_method == ResampleMethod.NP_MEAN:
-            return np.mean(subarrays, axis=0).astype(tile.dtype)
-        elif resample_method == ResampleMethod.NP_MEDIAN:
-            return np.median(subarrays, axis=0).astype(tile.dtype)
-        elif resample_method == ResampleMethod.NP_MAX:
-            return np.max(subarrays, axis=0).astype(tile.dtype)
-        elif resample_method == ResampleMethod.NP_MIN:
-            return np.min(subarrays, axis=0).astype(tile.dtype)
-        elif resample_method == ResampleMethod.NP_MAX_COLOR:
-            summed = np.sum(subarrays, axis=3)
-            pixel_selection = np.argmax(summed, axis=0)
-        elif resample_method == ResampleMethod.NP_MIN_COLOR:
-            summed = np.sum(subarrays, axis=3)
-            pixel_selection = np.argmin(summed, axis=0)
-        elif resample_method == ResampleMethod.NP_MODE:
-            # if a pixel occurs twice in a set of four, it is a mode
-            # if no mode, default to pixel 0. check for minimal matches 1=2, 1=3, 2=3
-            pixel_selection = np.where(
-                (
-                    (subarrays[1] == subarrays[2]).all(axis=2) |
-                    (subarrays[1] == subarrays[3]).all(axis=2)
-                ),
-                1, np.where(
-                    (subarrays[2] == subarrays[3]).all(axis=2),
-                    2, 0,
-                ),
-            )
-
-        if pixel_selection is not None:
-            if len(tile.shape) > 2:
-                pixel_selection = np.expand_dims(pixel_selection, axis=2)
-                pixel_selection = np.repeat(pixel_selection, tile.shape[2], axis=2)
-            return np.choose(pixel_selection, subarrays).astype(tile.dtype)
-        else:
-            msg = f'Unknown resample method {resample_method}.'
-            raise ValueError(msg)
+    if pixel_selection is not None:
+        if len(tile.shape) > 2:
+            pixel_selection = np.expand_dims(pixel_selection, axis=2)
+            pixel_selection = np.repeat(pixel_selection, tile.shape[2], axis=2)
+        return np.choose(pixel_selection, subarrays).astype(tile.dtype)
+    msg = f'Unknown resample method {resample_method}.'
+    raise ValueError(msg)
 
 
 def downsampleTileHalfRes(
@@ -123,7 +121,5 @@ def downsampleTileHalfRes(
                     resample_method,
                 )
             return result
-        else:
-            return pilResize(tile, new_shape, resample_method)
-    else:
-        return numpyResize(tile, new_shape, resample_method)
+        return pilResize(tile, new_shape, resample_method)
+    return numpyResize(tile, new_shape, resample_method)
