@@ -44,12 +44,9 @@ class TileIterator:
     def __next__(self) -> LazyTileDict:
         if self._iter is None:
             raise StopIteration
-        try:
-            tile = next(self._iter)
-            tile.setFormat(self.format, bool(self.resample), self._kwargs)
-            return tile
-        except StopIteration:
-            raise
+        tile = next(self._iter)
+        tile.setFormat(self.format, bool(self.resample), self._kwargs)
+        return tile
 
     def __repr__(self) -> str:
         repr = f'TileIterator<{self.source}'
@@ -67,6 +64,15 @@ class TileIterator:
         if self.info:
             return self.info
         return {}
+
+    def __len__(self) -> Optional[int]:
+        if self.info is None:
+            return None
+        iterlen = ((cast(int, self.info['xmax']) - cast(int, self.info['xmin'])) *
+                   (cast(int, self.info['ymax']) - cast(int, self.info['ymin'])))
+        if self.info.get('tile_position') is not None:
+            return 1 if cast(int, self.info['tile_position']) < iterlen else 0
+        return iterlen
 
     def _tileIteratorInfo(self, **kwargs) -> Optional[Dict[str, Any]]:  # noqa
         """
