@@ -213,28 +213,30 @@ def _imageToPIL(
             mode = modesBySize[image.shape[2] - 1]
         if len(image.shape) == 3 and image.shape[2] == 1:
             image = np.resize(image, image.shape[:2])
-        if image.dtype == np.uint32:
+        if cast(np.ndarray, image).dtype == np.uint32:
             image = np.floor_divide(image, 2 ** 24).astype(np.uint8)
-        elif image.dtype == np.uint16:
+        elif cast(np.ndarray, image).dtype == np.uint16:
             image = np.floor_divide(image, 256).astype(np.uint8)
-        elif image.dtype == np.int8:
-            image = (image.astype(float) + 128).astype(np.uint8)
-        elif image.dtype == np.int16:
-            image = np.floor_divide(image.astype(float) + 2 ** 15, 256).astype(np.uint8)
-        elif image.dtype == np.int32:
-            image = np.floor_divide(image.astype(float) + 2 ** 31, 2 ** 24).astype(np.uint8)
+        elif cast(np.ndarray, image).dtype == np.int8:
+            image = (cast(np.ndarray, image).astype(float) + 128).astype(np.uint8)
+        elif cast(np.ndarray, image).dtype == np.int16:
+            image = np.floor_divide(
+                cast(np.ndarray, image).astype(float) + 2 ** 15, 256).astype(np.uint8)
+        elif cast(np.ndarray, image).dtype == np.int32:
+            image = np.floor_divide(
+                cast(np.ndarray, image).astype(float) + 2 ** 31, 2 ** 24).astype(np.uint8)
         # TODO: The scaling of float data needs to be identical across all
         # tiles of an image.  This means that we need a reference to the parent
         # tile source or some other way of regulating it.
-        # elif image.dtype.kind == 'f':
+        # elif cast(np.ndarray, image).dtype.kind == 'f':
         #     if numpy.max(image) > 1:
         #         maxl2 = math.ceil(math.log(numpy.max(image) + 1) / math.log(2))
         #         image = image / ((2 ** maxl2) - 1)
         #     image = (image * 255).astype(numpy.uint8)
-        elif image.dtype != np.uint8:
+        elif cast(np.ndarray, image).dtype != np.uint8:
             image = np.clip(np.nan_to_num(np.where(
                 image is None, np.nan, image), nan=0), 0, 255).astype(np.uint8)
-        image = PIL.Image.fromarray(image, mode)
+        image = PIL.Image.fromarray(cast(np.ndarray, image), mode)
     elif not isinstance(image, PIL.Image.Image):
         image = PIL.Image.open(io.BytesIO(image))
     if setMode is not None and image.mode != setMode:
@@ -276,11 +278,11 @@ def _imageToNumpy(
         if len(image.shape) == 3:
             mode = modesBySize[(image.shape[2] - 1) if image.shape[2] <= 4 else 3]
             return image, mode
-        else:
-            mode = 'L'
-    if len(image.shape) == 2:
-        image = np.resize(image, (image.shape[0], image.shape[1], 1))
-    return image, mode
+        mode = 'L'
+    if len(cast(np.ndarray, image).shape) == 2:
+        image = np.resize(cast(np.ndarray, image),
+                          (cast(np.ndarray, image).shape[0], cast(np.ndarray, image).shape[1], 1))
+    return cast(np.ndarray, image), mode
 
 
 def _letterboxImage(image: PIL.Image.Image, width: int, height: int, fill: str) -> PIL.Image.Image:
@@ -310,7 +312,8 @@ def _letterboxImage(image: PIL.Image.Image, width: int, height: int, fill: str) 
     return result
 
 
-def _vipsCast(image: Any, mustBe8Bit: bool = False) -> Any:
+def _vipsCast(image: Any, mustBe8Bit: bool = False,
+              preferredCast: Optional[Tuple[Any, float, float]] = None) -> Any:
     """
     Cast a vips image to a format we want.
 
@@ -321,7 +324,7 @@ def _vipsCast(image: Any, mustBe8Bit: bool = False) -> Any:
     import pyvips
 
     image = cast(pyvips.Image, image)
-    formats = {
+    formats: Dict[pyvips.BandFormat, Tuple[pyvips.BandFormat, float, float]] = {
         pyvips.BandFormat.CHAR: (pyvips.BandFormat.UCHAR, 2**7, 1),
         pyvips.BandFormat.COMPLEX: (pyvips.BandFormat.USHORT, 0, 65535),
         pyvips.BandFormat.DOUBLE: (pyvips.BandFormat.USHORT, 0, 65535),
@@ -332,17 +335,21 @@ def _vipsCast(image: Any, mustBe8Bit: bool = False) -> Any:
         pyvips.BandFormat.SHORT: (pyvips.BandFormat.USHORT, 2**15, 1),
         pyvips.BandFormat.UINT: (pyvips.BandFormat.USHORT, 0, 2**-16),
     }
-    if image.format not in formats or (image.format == pyvips.BandFormat.USHORT and not mustBe8Bit):
+    if (image.format not in formats and preferredCast is None) or (
+            image.format == pyvips.BandFormat.USHORT and not mustBe8Bit):
         return image
-    target, offset, multiplier = formats[image.format]
-    if image.format == pyvips.BandFormat.DOUBLE or image.format == pyvips.BandFormat.FLOAT:
-        maxVal = image.max()
-        # These thresholds are higher than 256 and 65536 because bicubic and
-        # other interpolations can cause value spikes
-        if maxVal >= 2 and maxVal < 2**9:
-            multiplier = 256
-        elif maxVal >= 256 and maxVal < 2**17:
-            multiplier = 1
+    if preferredCast is not None:
+        target, offset, multiplier = preferredCast
+    else:
+        target, offset, multiplier = formats[image.format]
+        if image.format in {pyvips.BandFormat.DOUBLE, pyvips.BandFormat.FLOAT}:
+            maxVal = image.max()
+            # These thresholds are higher than 256 and 65536 because bicubic and
+            # other interpolations can cause value spikes
+            if maxVal >= 2 and maxVal < 2**9:
+                multiplier = 256
+            elif maxVal >= 256 and maxVal < 2**17:
+                multiplier = 1
     if mustBe8Bit and target != pyvips.BandFormat.UCHAR:
         target = pyvips.BandFormat.UCHAR
         multiplier /= 256
@@ -388,7 +395,8 @@ def _rasterioParameters(
 
     # add the remaining options
     options.update(tiled=True, bigtiff='IF_SAFER')
-    'predictor' not in options or options.update(predictor=predictor[str(options['predictor'])])
+    if 'predictor' in options:
+        options.update(predictor=predictor[str(options['predictor'])])
 
     return options
 
@@ -569,9 +577,9 @@ def dictToEtree(
             dictToEtree(v, elem)
         else:
             if k == 'text':
-                root.text = v
+                root.text = str(v)
             else:
-                root.set(k, v)
+                root.set(k, str(v))
     return root
 
 
@@ -618,7 +626,40 @@ def _arrayToPalette(palette: List[Union[str, float, Tuple[float, ...]]]) -> np.n
     return np.array(arr)
 
 
-def getPaletteColors(value: Union[str, List[Union[str, float, Tuple[float, ...]]]]) -> np.ndarray:
+def _mpl_lsc_to_palette(cmap: Any) -> List[str]:
+    """
+    Convert a matplotlib colormap to a palette of hexcolors.
+
+    :param cmap: a matplotlib LinearSegmentedColormap or ListedColormap.
+    :return: a list with hexadecimal color numbers.
+    """
+    import matplotlib as mpl
+
+    try:
+        if isinstance(cmap, mpl.colors.LinearSegmentedColormap):
+            ri = cast(Any, cmap)._segmentdata['red']
+            gi = cast(Any, cmap)._segmentdata['green']
+            bi = cast(Any, cmap)._segmentdata['blue']
+            ai = cast(Any, cmap)._segmentdata.get('alpha', None)
+            pal: List[str] = []
+            for idx in range(len(ri)):
+                r = int(round(float(ri[idx][-1]) * 255))
+                g = int(round(float(gi[idx][-1]) * 255))
+                b = int(round(float(bi[idx][-1]) * 255))
+                if ai is not None:
+                    a = int(round(float(ai[idx][-1]) * 255))
+                    entry = f'#{r:02X}{g:02X}{b:02X}{a:02X}'
+                else:
+                    entry = f'#{r:02X}{g:02X}{b:02X}'
+                if not len(pal) or pal[-1] != entry:
+                    pal.append(entry)
+            return pal
+    except Exception:
+        pass
+    return [mpl.colors.to_hex(cmap(i)) for i in range(cmap.N)]
+
+
+def getPaletteColors(value: Union[str, List[Union[str, float, Tuple[float, ...]]]]) -> np.ndarray:  # noqa
     """
     Given a list or a name, return a list of colors in the form of a numpy
     array of RGBA.  If a list, each entry is a color name resolvable by either
@@ -659,9 +700,21 @@ def getPaletteColors(value: Union[str, List[Union[str, float, Tuple[float, ...]]
                 cmap = (mpl.colormaps.get_cmap(str(value)) if hasattr(getattr(
                     mpl, 'colormaps', None), 'get_cmap') else
                     mpl.cm.get_cmap(str(value)))
-                palette = [mpl.colors.to_hex(cmap(i)) for i in range(cmap.N)]
+                palette = _mpl_lsc_to_palette(cmap)  # type: ignore
         except (ImportError, ValueError, AttributeError):
             pass
+    if palette is None:
+        if str(value).startswith('tol.'):
+            key = value[4:]
+            try:
+                import tol_colors
+
+                if key in tol_colors.colorsets:
+                    palette = list(tol_colors.colorsets[key])
+                elif key in tol_colors.TOLcmaps().namelist:
+                    palette = _mpl_lsc_to_palette(tol_colors.tol_cmap(key))  # type: ignore
+            except ImportError:
+                pass
     if palette is None:
         raise ValueError('cannot be used as a color palette.: %r.' % value)
     return _arrayToPalette(palette)
@@ -734,6 +787,13 @@ def getAvailableNamedPalettes(includeColors: bool = True, reduced: bool = False)
                 palettes.add(key)
     except ImportError:
         pass
+    try:
+        import tol_colors
+
+        palettes |= {f'tol.{key}' for key in tol_colors.colorsets}
+        palettes |= {f'tol.{key}' for key in tol_colors.TOLcmaps().namelist}
+    except ImportError:
+        pass
     if reduced:
         palettes = {
             key for key in palettes
@@ -758,6 +818,12 @@ def fullAlphaValue(arr: Union[np.ndarray, npt.DTypeLike]) -> int:
         dtype = np.dtype(dtype)
     if cast(np.dtype, dtype).kind == 'u':
         return np.iinfo(dtype).max
+    if isinstance(arr, np.ndarray) and cast(np.dtype, dtype).kind == 'f':
+        amax = cast(float, np.amax(arr))
+        if amax > 1 and amax < 256:
+            return 255
+        if amax > 1 and amax < 65536:
+            return 65535
     return 1
 
 
@@ -786,24 +852,24 @@ def _makeSameChannelDepth(arr1: np.ndarray, arr2: np.ndarray) -> Tuple[np.ndarra
     # If any array is RGB, make sure all arrays are RGB.
     for key, arr in arrays.items():
         other = arrays['arr1' if key == 'arr2' else 'arr2']
-        if arr.shape[2] < 3 and other.shape[2] >= 3:  # type: ignore[misc]
+        if arr.shape[2] < 3 and other.shape[2] >= 3:
             newarr = np.ones(
-                (arr.shape[0], arr.shape[1], arr.shape[2] + 2),  # type: ignore[misc]
+                (arr.shape[0], arr.shape[1], arr.shape[2] + 2),
                 dtype=arr.dtype)
             newarr[:, :, 0:1] = arr[:, :, 0:1]
             newarr[:, :, 1:2] = arr[:, :, 0:1]
             newarr[:, :, 2:3] = arr[:, :, 0:1]
-            if arr.shape[2] == 2:  # type: ignore[misc]
+            if arr.shape[2] == 2:
                 newarr[:, :, 3:4] = arr[:, :, 1:2]
             arrays[key] = newarr
     # If only one array has an A channel, make sure all arrays have an A
     # channel
     for key, arr in arrays.items():
         other = arrays['arr1' if key == 'arr2' else 'arr2']
-        if arr.shape[2] < other.shape[2]:  # type: ignore[misc]
+        if arr.shape[2] < other.shape[2]:
             arrays[key] = np.pad(
                 arr,
-                ((0, 0), (0, 0), (0, other.shape[2] - arr.shape[2])),  # type: ignore[misc]
+                ((0, 0), (0, 0), (0, other.shape[2] - arr.shape[2])),
                 constant_values=fullAlphaValue(arr))
     return arrays['arr1'], arrays['arr2']
 
@@ -827,8 +893,8 @@ def _addSubimageToImage(
     if image is None:
         if (x, y, width, height) == (0, 0, subimage.shape[1], subimage.shape[0]):
             return subimage
-        image = np.empty(
-            (height, width, subimage.shape[2]),  # type: ignore[misc]
+        image = np.zeros(
+            (height, width, subimage.shape[2]),
             dtype=subimage.dtype)
     elif len(image.shape) != len(subimage.shape) or image.shape[-1] != subimage.shape[-1]:
         image, subimage = _makeSameChannelDepth(image, subimage)
@@ -886,7 +952,7 @@ def _addRegionTileToTiled(
         subimage = subimage.astype('d')
     vimgMem = pyvips.Image.new_from_memory(
         np.ascontiguousarray(subimage).data,
-        subimage.shape[1], subimage.shape[0], subimage.shape[2],  # type: ignore[misc]
+        subimage.shape[1], subimage.shape[0], subimage.shape[2],
         dtypeToGValue[subimage.dtype.char])
     vimg = pyvips.Image.new_temp_file('%s.v')
     vimgMem.write(vimg)
@@ -897,7 +963,7 @@ def _addRegionTileToTiled(
             'mm_x': tile.get('mm_x') if tile else None,
             'mm_y': tile.get('mm_y') if tile else None,
             'magnification': tile.get('magnification') if tile else None,
-            'channels': subimage.shape[2],  # type: ignore[misc]
+            'channels': subimage.shape[2],
             'strips': {},
         }
     if y not in image['strips']:
@@ -939,7 +1005,7 @@ def _calculateWidthHeight(
     scaledHeight = max(1, int(regionHeight * cast(float, width) / regionWidth))
     if scaledWidth == width or (
             cast(float, width) * regionHeight > cast(float, height) * regionWidth and
-            not scaledHeight == height):
+            scaledHeight != height):
         scale = float(regionHeight) / cast(float, height)
         width = scaledWidth
     else:
