@@ -26,10 +26,11 @@ from importlib.metadata import version as _importlib_version
 import numpy as np
 import PIL.Image
 
-from large_image.cache_util import LruCacheMetaclass, methodcache
-from large_image.constants import (TILE_FORMAT_IMAGE, TILE_FORMAT_NUMPY,
-                                   TILE_FORMAT_PIL, SourcePriority,
-                                   TileOutputMimeTypes)
+from large_image import config
+from large_image.cache_util import LruCacheMetaclass, _cacheClearFuncs, methodcache
+from large_image.constants import (PROJECTION_SENTINEL, TILE_FORMAT_IMAGE,
+                                   TILE_FORMAT_NUMPY, TILE_FORMAT_PIL,
+                                   SourcePriority, TileOutputMimeTypes)
 from large_image.exceptions import (TileSourceError,
                                     TileSourceFileNotFoundError,
                                     TileSourceInefficientError)
@@ -48,6 +49,7 @@ gdal = None
 gdal_array = None
 gdalconst = None
 osr = None
+ogr = None
 pyproj = None
 
 
@@ -76,6 +78,15 @@ def _lazyImport():
             import pyproj
 
             # isort: on
+
+            def _clearGDALCache():
+                old = gdal.GetCacheMax()
+                # print('Clearing GDAL cache: size %r, max %r' % (
+                #     gdal.GetCacheUsed(), old))
+                gdal.SetCacheMax(0)
+                gdal.SetCacheMax(old)
+
+            _cacheClearFuncs.append(_clearGDALCache)
         except ImportError:
             msg = 'gdal module not found.'
             raise TileSourceError(msg)
@@ -89,7 +100,10 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
     cacheName = 'tilesource'
     name = 'gdal'
 
-    def __init__(self, path, projection=None, unitsPerPixel=None, **kwargs):
+    VECTOR_IMAGE_SIZE = 256 * 1024  # for vector files without projections
+    PROJECTED_VECTOR_IMAGE_SIZE = 32 * 1024  # if the file has a projection
+
+    def __init__(self, path, projection=PROJECTION_SENTINEL, unitsPerPixel=None, **kwargs):  # noqa
         """
         Initialize the tile class.  See the base class for other available
         parameters.
@@ -115,8 +129,10 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
         self._bounds = {}
         self._largeImagePath = self._getLargeImagePath()
         try:
-            self.dataset = gdal.Open(self._largeImagePath, gdalconst.GA_ReadOnly)
-        except (RuntimeError, UnicodeDecodeError):
+            self.dataset = gdal.OpenEx(self._largeImagePath, gdalconst.GA_ReadOnly)
+            if not self.dataset.RasterCount and self.dataset.GetLayer() is not None:
+                self.dataset = self._openVectorSource(self.dataset)
+        except (RuntimeError, UnicodeDecodeError, OverflowError):
             if not os.path.isfile(self._largeImagePath):
                 raise TileSourceFileNotFoundError(self._largeImagePath) from None
             msg = 'File cannot be opened via GDAL'
@@ -125,6 +141,11 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
         self.tileSize = 256
         self.tileWidth = self.tileSize
         self.tileHeight = self.tileSize
+        if projection == PROJECTION_SENTINEL:
+            if self.isGeospatial(self.dataset):
+                projection = config.getConfig('default_projection')
+            else:
+                projection = None
         if projection and projection.lower().startswith('epsg:'):
             projection = projection.lower()
         if projection and not isinstance(projection, bytes):
@@ -141,7 +162,7 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
         is_netcdf = self._checkNetCDF()
         try:
             scale = self.getPixelSizeInMeters()
-        except RuntimeError as exc:
+        except (RuntimeError, ZeroDivisionError) as exc:
             raise TileSourceError('File cannot be opened via GDAL: %r' % exc)
         if not self.sizeX or not self.sizeY:
             msg = 'File cannot be opened via GDAL (no size)'
@@ -161,6 +182,54 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
         self._getPopulatedLevels()
         self._getTileLock = threading.Lock()
         self._setDefaultStyle()
+
+    def _openVectorSource(self, vec):
+        # This is mainly to speed up back-to-back canRead and open calls
+        if hasattr(self.__class__, '_openVectorLock') and hasattr(self.__class__, '_lastVectorDS'):
+            with self.__class__._openVectorLock:
+                if self.__class__._lastVectorDS[0] == self._largeImagePath:
+                    return self.__class__._lastVectorDS[1]
+        layer = vec.GetLayer()
+        x_min, x_max, y_min, y_max = layer.GetExtent()
+        try:
+            proj = layer.GetSpatialRef().ExportToWkt()
+            if (layer.GetSpatialRef().ExportToProj4() == '+proj=longlat +datum=WGS84 +no_defs' and
+                    (max(abs(x_min), abs(x_max)) > 180 or max(abs(y_min), abs(y_max)) > 90)):
+                proj = None
+        except Exception:
+            proj = None
+        self._geospatial = proj is not None
+        # Define raster parameters
+        pixel_size = max(x_max - x_min, y_max - y_min) / (
+            self.VECTOR_IMAGE_SIZE if proj is None else self.PROJECTED_VECTOR_IMAGE_SIZE)
+        if not pixel_size:
+            msg = 'Cannot determine dimensions'
+            raise RuntimeError(msg)
+        if not proj and pixel_size < 1:
+            pixel_size = 1
+        x_res = int((x_max - x_min) / pixel_size)
+        y_res = int((y_max - y_min) / pixel_size)
+
+        # Create an in-memory raster dataset
+        with tempfile.NamedTemporaryFile(suffix='.tif', delete=True) as tmpfile:
+            drv = gdal.GetDriverByName('Gtiff')
+            ds = drv.Create(
+                tmpfile.name, x_res, y_res, 1, gdal.GDT_Byte,
+                options=['COMPRESS=DEFLATE', 'PREDICTOR=2', 'TILED=YES',
+                         'SPARSE_OK=TRUE', 'BIGTIFF=IF_SAFER'])
+            # Set raster geotransform and projection
+            ds.SetGeoTransform((x_min, pixel_size, 0, y_min, 0, pixel_size))
+            if proj:
+                ds.SetProjection(proj)
+            msg = (f'Rasterizing a vector layer to {x_res} x {y_res} '
+                   f'({"not " if not self._geospatial else ""} geospatial)')
+            self.logger.info(msg)
+            gdal.RasterizeLayer(ds, [1], layer, burn_values=[255])
+            if not hasattr(self.__class__, '_openVectorLock'):
+                self.__class__._openVectorLock = threading.RLock()
+            with self.__class__._openVectorLock:
+                self.__class__._lastVectorDS = (self._largeImagePath, ds)
+            return ds
 
     def _getDriver(self):
         """
@@ -272,9 +341,11 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
 
     @staticmethod
     def getLRUHash(*args, **kwargs):
+        proj = kwargs.get('projection', args[1] if len(args) >= 2 else PROJECTION_SENTINEL)
+        proj = proj if proj != PROJECTION_SENTINEL else config.getConfig('default_projection')
         return super(GDALFileTileSource, GDALFileTileSource).getLRUHash(
             *args, **kwargs) + ',%s,%s' % (
-                kwargs.get('projection', args[1] if len(args) >= 2 else None),
+                proj,
                 kwargs.get('unitsPerPixel', args[3] if len(args) >= 4 else None))
 
     def getState(self):
@@ -296,7 +367,7 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
             if (self.dataset.GetGeoTransform(can_return_null=True) or
                     hasattr(self, '_netcdf') or self._getDriver() in {'NITF'}):
                 return 'epsg:4326'
-            return
+            return None
         proj = osr.SpatialReference()
         proj.ImportFromWkt(wkt)
         return proj.ExportToProj4()
@@ -328,7 +399,7 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
         if isinstance(proj, bytes):
             proj = proj.decode()
         if not isinstance(proj, str):
-            return
+            return None
         if proj.lower().startswith('proj4:'):
             proj = proj.split(':', 1)[1]
         if proj.lower().startswith('epsg:'):
@@ -382,7 +453,7 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
             'epsg:', or '+proj=' or a enumerated value like 'wgs84', or one of
             the super's values.
         :param kwargs: optional parameters.
-        :returns: left, top, right, bottom, units.  The new bounds in the
+        :returns: left, top, right, bottom, width, height, units.  The new bounds in the
             either pixel or class projection units.
         """
         if not kwargs.get('unitsWH') or kwargs.get('unitsWH') == units:
@@ -394,6 +465,11 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
                 top = bottom - height
             if bottom is None and top is not None and height is not None:
                 bottom = top + height
+        elif self.projection:
+            left, top, right, bottom, width, height, units = self._applyUnitsWH(
+                left, top, right, bottom, width, height, units, kwargs.get('unitsWH'),
+            )
+
         if (left is None and right is None) or (top is None and bottom is None):
             msg = ('Cannot convert from projection unless at least one of '
                    'left and right and at least one of top and bottom is '
@@ -424,7 +500,7 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
         top = ptop if top is not None else None
         right = pright if right is not None else None
         bottom = pbottom if bottom is not None else None
-        return left, top, right, bottom, units
+        return left, top, right, bottom, width, height, units
 
     def pixelToProjection(self, x, y, level=None):
         """
@@ -464,7 +540,7 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
             nativeSrs = self.getProj4String()
             if not nativeSrs or not gt:
                 self._bounds[srs] = None
-                return
+                return None
             bounds = {
                 'll': {
                     'x': gt[0] + self.sourceSizeY * gt[2],
@@ -598,11 +674,13 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
         """
         This is true if the source has geospatial information.
         """
-        return bool(
-            self.dataset.GetProjection() or
-            (self.dataset.GetGCPProjection() and self.dataset.GetGCPs()) or
-            self.dataset.GetGeoTransform(can_return_null=True) or
-            hasattr(self, '_netcdf'))
+        if not hasattr(self, '_geospatial'):
+            self._geospatial = bool(
+                self.dataset.GetProjection() or
+                (self.dataset.GetGCPProjection() and self.dataset.GetGCPs()) or
+                self.dataset.GetGeoTransform(can_return_null=True) or
+                hasattr(self, '_netcdf'))
+        return self._geospatial
 
     def getMetadata(self):
         metadata = super().getMetadata()
@@ -907,8 +985,13 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
             '-ts', str(int(math.floor(outWidth))), str(int(math.floor(outHeight))),
         ]
 
-        fd, outputPath = tempfile.mkstemp('.tiff', 'tiledGeoRegion_')
-        os.close(fd)
+        outputPath = kwargs.get('output', {}).get('path')
+        if outputPath is not None:
+            outputPath = pathlib.Path(outputPath)
+            outputPath.parent.mkdir(parents=True, exist_ok=True)
+        else:
+            fd, outputPath = tempfile.mkstemp('.tiff', 'tiledGeoRegion_')
+            os.close(fd)
         try:
             self.logger.info('Using gdal warp %r', gdalParams)
             ds = gdal.Open(self._largeImagePath, gdalconst.GA_ReadOnly)
@@ -921,7 +1004,7 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
             raise exc
         return pathlib.Path(outputPath), TileOutputMimeTypes['TILED']
 
-    def validateCOG(self, check_tiled=True, full_check=False, strict=True, warn=True):
+    def validateCOG(self, check_tiled=True, full_check=False, strict=True, warn=True) -> bool:
         """Check if this image is a valid Cloud Optimized GeoTiff.
 
         This will raise a :class:`large_image.exceptions.TileSourceInefficientError`
@@ -945,8 +1028,8 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
         """
         from osgeo_utils.samples.validate_cloud_optimized_geotiff import validate
 
-        warnings, errors, details = validate(
-            self._largeImagePath,
+        warnings, errors, _ = validate(
+            self.dataset,
             check_tiled=check_tiled,
             full_check=full_check,
         )
@@ -960,19 +1043,20 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
         return True
 
     @staticmethod
-    def isGeospatial(path):
+    def isGeospatial(ds):
         """
-        Check if a path is likely to be a geospatial file.
+        Check if a GDAL Dataset or file path is likely to be geospatial.
 
-        :param path: The path to the file
+        :param ds: A GDAL Dataset or the path to the file
         :returns: True if geospatial.
         """
         _lazyImport()
 
-        try:
-            ds = gdal.Open(str(path), gdalconst.GA_ReadOnly)
-        except Exception:
-            return False
+        if not isinstance(ds, gdal.Dataset):
+            try:
+                ds = gdal.Open(str(ds), gdalconst.GA_ReadOnly)
+            except Exception:
+                return False
         if ds:
             if ds.GetGCPs() and ds.GetGCPProjection():
                 return True
@@ -1000,7 +1084,7 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
             cls.mimeTypes = cls.mimeTypes.copy()
             for idx in range(gdal.GetDriverCount()):
                 drv = gdal.GetDriver(idx)
-                if drv.GetMetadataItem(gdal.DCAP_RASTER):
+                if drv.GetMetadataItem(gdal.DCAP_RASTER) or drv.GetMetadataItem(gdal.DCAP_VECTOR):
                     drvexts = drv.GetMetadataItem(gdal.DMD_EXTENSIONS)
                     if drvexts is not None:
                         for ext in drvexts.split():
@@ -1010,6 +1094,16 @@ class GDALFileTileSource(GDALBaseFileTileSource, metaclass=LruCacheMetaclass):
                     if drvmimes is not None:
                         if drvmimes not in cls.mimeTypes:
                             cls.mimeTypes[drvmimes] = SourcePriority.IMPLICIT
+            # This list was compiled by trying to read the test files in GDAL's
+            # repo.
+            for ext in {
+                    'adf', 'aux', 'demtif', 'dim', 'doq', 'flt', 'fst',
+                    'gdbtable', 'gsc', 'h3', 'idf', 'lan', 'los', 'lrc',
+                    'mapml', 'mint.bin', 'mtw', 'nsf', 'nws', 'on2', 'on9',
+                    'osm.pbf', 'pjg', 'prj', 'ptf', 'rasterlite', 'rdb', 'rl2',
+                    'shx', 'sos', 'tif.grd', 'til', 'vic', 'xlb'}:
+                if ext.lower() not in cls.extensions:
+                    cls.extensions[ext.lower()] = SourcePriority.IMPLICIT_LOW
 
 
 def open(*args, **kwargs):
