@@ -27,9 +27,9 @@ registry = {
     # RGB JPEG compression
     # Source: TCGA-AA-A02O-11A-01-BS1.8b76f05c-4a8b-44ba-b581-6b8b4f437367.svs
     'TCGA-AA-A02O-11A-01-BS1.8b76f05c-4a8b-44ba-b581-6b8b4f437367.svs': 'sha512:1b75a4ec911017aef5c885760a3c6575dacf5f8efb59fb0e011108dce85b1f4e97b8d358f3363c1f5ea6f1c3698f037554aec1620bbdd4cac54e3d5c9c1da1fd',  # noqa
-    # Tiff with floating point pixels
+    # Tiff with floating point pixels (icc profile stripped)
     # Source: d042-353.crop.small.float32.tif
-    'd042-353.crop.small.float32.tif': 'sha512:ae05dbe6f3330c912893b203b55db27b0fdf3222a0e7f626d372c09668334494d07dc1d35533670cfac51b588d2292eeee7431317741fdb4cbb281c28a289115',  # noqa
+    'd042-353.crop.small.float32.tif': 'sha512:8b640e9adcd0b8aba794666027b80215964d075e76ca2ebebefc7e17c3cd79af7da40a40151e2a2ba0ae48969e54275cf69a3cfc1a2a6b87fbb0d186013e5489',  # noqa
     # JPEG with progressive compression and restart markers
     # Source: d042-353.crop.small.jpg
     'd042-353.crop.small.jpg': 'sha512:1353646637c1fae266b87312698aa39eca0311222c3a1399b60efdc13bfe55e2f3db59da005da945dd7e9e816f31ccd18846dd72744faac75215074c3d87414f',  # noqa
@@ -113,7 +113,9 @@ registry = {
     '18f6378f-433c-42bf-9373-1ff9c808c118.dcm': 'sha512:36432183380eb7d44417a2210a19d550527abd1181255e19ed5c1d17695d8bb8ca42f5b426a63fa73b84e0e17b770401a377ae0c705d0ed7fdf30d571ef60e2d',  # noqa
     'a131592c-a069-4aa7-8031-398654aa8a3d.dcm': 'sha512:99bd3da4b8e11ce7b4f7ed8a294ed0c37437320667a06c40c383f4b29be85fe8e6094043e0600bee0ba879f2401de4c57285800a4a23da2caf2eb94e5b847ee0',  # noqa
     # Synthetic newer ndpi with binary data and nonblank image labelled as RGB
-    'synthetic_ndpi_2024.ndpi': 'sha512:192cdcf551a824277ef70358b8ed6225dca0c5b5d0817fe0d800b72638e0ad9416cd5bc40cb186219da60fd324b676d1b32cc05a276a33b71b485d665a31e42e',  # noqa
+    'synthetic_ndpi_2025.ndpi': 'sha512:b9b2c420cde9fd988786afc02efb761a7d425ce542cc68f10f5878bdc7177d61952e0d52508501c82d5664a07a87e7486ec4bb0b6634c556400cfc91fc3f52ec',  # noqa
+    # Synthetic ndpi with multiple focal planes
+    'synthetic_ndpi_multiplane_2025.ndpi': 'sha512:1025d6ddd74070d0bb2c3ab398bc5e6ae05390651f84df22c37e3dbe547bb16b4eba0a0a3cefc7ac824ff2e87320782b676b2049a602bc0d5dbb105bbc94e888',  # noqa
     # Synthetic uint16 untiled tiff that can be read with the tiff source
     'synthetic_untiled_16.tiff': 'sha512:f4773fcfa749ba9c2db25319c9e8ad8586dd148de4366dae0393a3703906dace9f11233eafdb24418b598170d6372ef1ca861bf8d7a8212cac21a0eb8636ee77',  # noqa
     # DICOM with int16 data
@@ -123,6 +125,10 @@ registry = {
     # Synthetic Indica Labs tiff; subifds missing tile/strip data and unmarked
     # float32 pixels rather than uint32
     'synthetic_indica.tiff': 'sha512:fba7eb2fb5fd12ac242d8b0760440f170f48f9e2434a672cbf230bd8a9ff02fad8f9bdf7225edf2de244f412edfc5205e695031a1d43dd99fe31c3aca11909a1',  # noqa
+    # Converted from the TCGA svs file using bioformats java program and
+    # --rgb --quality=0.015 --compression='JPEG-2000 Lossy' parameters to make
+    # the file small
+    'TCGA-55-8207-01Z-00-DX1.ome.tiff': 'sha512:50cf63f0e8bfa3054d3532b7dd0237b66aeb4c7609da874639a28bc068dbd157f786e84d3eb76a3b0e6636a042c56c3b96d3be2ad66f7589d0542a5d20cecdb4',  # noqa
 }
 
 
