@@ -1,4 +1,5 @@
 import subprocess
+from datetime import datetime, timezone
 from multiprocessing.pool import Pool, ThreadPool
 from os import sys
 
@@ -10,6 +11,7 @@ from PIL import Image
 
 import large_image
 from large_image.constants import NEW_IMAGE_PATH_FLAG
+from large_image.exceptions import TileSourceError
 from large_image.tilesource.resample import ResampleMethod
 
 TMP_DIR = 'tmp/zarr_sink'
@@ -93,12 +95,44 @@ def testAddTileWithLevel():
     assert arrays.get('1') is None
 
 
+def testNoSamplesAxis(tmp_path):
+    output_file = tmp_path / 'test.tiff'
+    sink = large_image_source_zarr.new()
+    data = np.zeros((4, 1040, 1388))
+    sink.addTile(data, 0, 0, axes=['c', 'y', 'x'])
+    sink.write(output_file)
+    assert sink.metadata.get('bandCount') == 1
+
+
 def testExtraAxis():
     sink = large_image_source_zarr.new()
     sink.addTile(np.random.random((256, 256)), 0, 0, z=1)
     metadata = sink.getMetadata()
     assert metadata.get('bandCount') == 1
     assert len(metadata.get('frames')) == 2
+
+
+def testXYAxis():
+    sink = large_image_source_zarr.new()
+    sink.addTile(np.random.random((256, 256)), 0, 0, xy=1)
+    metadata = sink.getMetadata()
+    assert metadata['IndexStride']['IndexXY'] == 1
+
+
+def testXLength1():
+    sink = large_image_source_zarr.new()
+    sink.addTile(np.zeros((1, 1, 1)), x=0, y=138)
+    sink.addTile(np.zeros((1, 1, 1)), x=0, y=138)
+    metadata = sink.getMetadata()
+    assert metadata.get('sizeX') == 1
+
+
+def testMultiFrameAxes():
+    sink = large_image_source_zarr.new()
+    sink.addTile(np.random.random((256, 256)), 0, 0, q=1)
+    assert sink.metadata.get('IndexRange') == dict(IndexQ=2)
+    sink.addTile(np.random.random((256, 256)), 0, 0, r=1)
+    assert sink.metadata.get('IndexRange') == dict(IndexQ=2, IndexR=2)
 
 
 @pytest.mark.parametrize('file_type', FILE_TYPES)
@@ -561,11 +595,11 @@ def get_expected_metadata(axis_spec, frame_shape):
         },
         **{
             f'Value{k.upper()}': dict(
-                values=v.get('all_values', v['values']),
+                values=v['values'],
                 units=v['units'],
                 uniform=v['uniform'],
-                min=min(v.get('all_values', v['values'])),
-                max=max(v.get('all_values', v['values'])),
+                min=min(v['values']),
+                max=max(v['values']),
                 datatype=v['dtype'],
             ) for k, v in axis_spec.items()
         },
@@ -600,10 +634,6 @@ def testFrameValuesSmall(use_add_tile_args, tmp_path):
     ))
     expected_metadata = get_expected_metadata(axis_spec, frame_shape)
 
-    sink.frameAxes = list(axis_spec.keys())
-    sink.frameUnits = {
-        k: v['units'] for k, v in axis_spec.items()
-    }
     frame_values_shape = [
         *[len(v['values']) for v in axis_spec.values()],
         len(axis_spec),
@@ -633,7 +663,11 @@ def testFrameValuesSmall(use_add_tile_args, tmp_path):
     index += 1
 
     if not use_add_tile_args:
+        sink.frameAxes = list(axis_spec.keys())
         sink.frameValues = frame_values
+    sink.frameUnits = {
+        k: v['units'] for k, v in axis_spec.items()
+    }
     compare_metadata(dict(sink.getMetadata()), expected_metadata)
 
     sink.write(output_file)
@@ -657,20 +691,6 @@ def testFrameValues(use_add_tile_args, tmp_path):
         ),
         t=dict(
             values=[10.0, 20.0, 30.0],
-            all_values=(
-                [10.00] * 3 +
-                [20.00] * 3 +
-                [30.00] * 3 +
-                [10.01] * 3 +
-                [20.01] * 3 +
-                [30.01] * 3 +
-                [10.02] * 3 +
-                [20.02] * 3 +
-                [30.02] * 3 +
-                [10.03] * 3 +
-                [20.03] * 3 +
-                [30.03] * 3
-            ),
             uniform=False,
             units='millisecond',
             stride=3,
@@ -686,10 +706,6 @@ def testFrameValues(use_add_tile_args, tmp_path):
     )
     expected_metadata = get_expected_metadata(axis_spec, frame_shape)
 
-    sink.frameAxes = list(axis_spec.keys())
-    sink.frameUnits = {
-        k: v['units'] for k, v in axis_spec.items()
-    }
     frame_values_shape = [
         *[len(v['values']) for v in axis_spec.values()],
         len(axis_spec),
@@ -727,12 +743,17 @@ def testFrameValues(use_add_tile_args, tmp_path):
             index += 1
 
     if not use_add_tile_args:
+        sink.frameAxes = list(axis_spec.keys())
         sink.frameValues = frame_values
+    sink.frameUnits = {
+        k: v['units'] for k, v in axis_spec.items()
+    }
     compare_metadata(dict(sink.getMetadata()), expected_metadata)
 
     sink.write(output_file)
     written = large_image_source_zarr.open(output_file)
-    compare_metadata(dict(written.getMetadata()), expected_metadata)
+    assert written.metadata['IndexRange'] == expected_metadata['IndexRange']
+    assert written.metadata['IndexStride']['IndexC'] == 1
 
 
 def testFrameValuesEdgeCases(tmp_path):
@@ -783,19 +804,305 @@ def testSubprocess(tmp_path):
     subprocess.run([sys.executable, '-c', """import large_image_source_zarr
 import numpy as np
 sink = large_image_source_zarr.open('%s')
-sink.addTile(np.ones((1, 1, 1)), x=2047, y=2047, t=5, z=2)
+sink.addTile(np.ones((1, 1, 1)), x=2047, y=2047, t=5, z=2, t_value='thursday', z_value=0.2)
 """ % path], capture_output=True, text=True, check=True)
-    sink.addTile(np.ones((1, 1, 1)), x=5000, y=4095, t=0, z=4)
+    sink.addTile(np.ones((1, 1, 1)), x=5000, y=4095, t=0, z=4, t_value='sunday', z_value=0.4)
 
-    assert sink.metadata['IndexRange']['IndexZ'] == 5
+    metadata = sink.getMetadata()
+    assert metadata['IndexRange']['IndexZ'] == 5
     assert sink.getRegion(
         region=dict(left=2047, top=2047, width=1, height=1),
         format='numpy',
         frame=17,
     )[0] == 1
+    assert metadata['frames'][17]['ValueT'] == 'thursday'
+    assert metadata['frames'][17]['ValueZ'] == 0.2
     assert sink.getRegion(
         region=dict(left=5000, top=4095, width=1, height=1),
         format='numpy',
         frame=24,
     )[0] == 1
+    assert metadata['frames'][24]['ValueT'] == 'sunday'
+    assert metadata['frames'][24]['ValueZ'] == 0.4
     assert sink.sizeX == 5001
+
+
+@pytest.mark.parametrize('axes_order', ['tzd', 'tdz', 'dzt', 'dtz', 'ztd', 'zdt'])
+def testAddAxes(tmp_path, axes_order):
+    sink = large_image_source_zarr.new()
+    kwarg_groups = [
+        dict(t=0, t_value='sunday'),
+        dict(
+            t=5, t_value='friday',
+            z=1, z_value=0.1,
+            axes=axes_order.replace('d', '') + 'yxs',
+        ),
+        dict(
+            t=6, t_value='saturday',
+            z=2, z_value=0.2,
+            d=1, d_value=100,
+            axes=axes_order + 'yxs',
+        ),
+    ]
+    for kwarg_group in kwarg_groups:
+        sink.addTile(
+            np.ones((4, 4, 4)),
+            x=1020, y=1020,
+            **kwarg_group,
+        )
+
+    metadata = sink.getMetadata()
+    t_stride = metadata['IndexStride']['IndexT']
+    z_stride = metadata['IndexStride']['IndexZ']
+    expected_filled_frames = [
+        # first and last frame are known, middle frame depends on axis ordering
+        0, z_stride + t_stride * 5, 41,
+    ]
+    for frame in metadata.get('frames', []):
+        frame_index = frame.get('Frame')
+        sample = sink.getRegion(
+            region=dict(left=1020, top=1020, width=1, height=1),
+            format='numpy',
+            frame=frame_index,
+        )[0]
+        frame_values = dict(
+            t_value=frame.get('ValueT'),
+            z_value=frame.get('ValueZ'),
+            d_value=frame.get('ValueD'),
+        )
+        kwarg_group = {}
+        if frame_index in expected_filled_frames:
+            kwarg_group = kwarg_groups[expected_filled_frames.index(frame_index)]
+            assert (sample == 1).all()
+        else:
+            assert (sample == 0).all()
+
+        for k, v in frame_values.items():
+            assert v == kwarg_group.get(k, 0)
+
+
+def testMinWidthMinHeight(tmp_path):
+    output_file = tmp_path / 'test.db'
+    sink = large_image_source_zarr.new()
+    sink.addTile(np.zeros((256, 256, 1), dtype=np.uint8), x=0, y=0)
+    sink.minWidth = 1024
+    sink.minHeight = 2048
+    sink.addTile(np.zeros((256, 256, 1), dtype=np.uint8), x=256, y=0)
+    sink.minWidth = 768
+    sink.minHeight = 800
+    sink.write(output_file)
+
+    written = large_image_source_zarr.open(output_file)
+    metadata = written.getMetadata()
+    assert metadata.get('sizeX') == 768
+    assert metadata.get('sizeY') == 800
+
+
+def testNegativeMinWidth():
+    sink = large_image_source_zarr.new()
+    with pytest.raises(TileSourceError) as e:
+        sink.minWidth = -10
+    assert str(e.value) == 'minWidth must be positive or None'
+
+
+def testNegativeMinHeight():
+    sink = large_image_source_zarr.new()
+    with pytest.raises(TileSourceError) as e:
+        sink.minHeight = -10
+    assert str(e.value) == 'minHeight must be positive or None'
+
+
+def testDescriptionAndAdditionalMetadata(tmp_path):
+    output_file = tmp_path / 'test.tiff'
+    sink = large_image_source_zarr.new()
+    sink.addTile(np.zeros((256, 256, 1), dtype=np.uint8), x=0, y=0)
+
+    description = 'This is a test description.'
+    additional_metadata = dict(
+        name='Test',
+        values=[1, 2, 3],
+        nested=dict(hello='world'),
+    )
+    both = dict(
+        description=description,
+        additionalMetadata=additional_metadata,
+    )
+
+    sink.imageDescription = description
+    assert sink._imageDescription == description
+    sink.additionalMetadata = additional_metadata
+    assert sink._imageDescription == both
+    assert sink.imageDescription == description
+    assert sink.additionalMetadata == additional_metadata
+
+    # modify values and check again
+    description = 'This is another test description'
+    both['description'] = description
+    additional_metadata['name'] = 'Test 2'
+    both['additionalMetadata'] = additional_metadata
+    sink.imageDescription = description
+    sink.additionalMetadata = additional_metadata
+    assert sink._imageDescription == both
+    assert sink.imageDescription == description
+    assert sink.additionalMetadata == additional_metadata
+
+    sink.write(output_file)
+    written = large_image.open(output_file)
+    internal = written.getInternalMetadata()['xml']['internal']['zarr']['base']
+    assert internal['multiscales'][0]['metadata']['description'] == both
+
+
+def testRehydrateDescriptionAndAdditionalMetadata(tmp_path):
+    output_file = tmp_path / 'test.db'
+    sink = large_image_source_zarr.new()
+    sink.addTile(np.zeros((256, 256, 1), dtype=np.uint8), x=0, y=0)
+
+    description = 'This is a test description.'
+    additional_metadata = dict(
+        name='Test',
+        values=[1, 2, 3],
+        nested=dict(hello='world'),
+    )
+    sink.imageDescription = description
+    sink.additionalMetadata = additional_metadata
+
+    sink.write(output_file)
+    written = large_image_source_zarr.open(output_file)
+    assert written.imageDescription == description
+    assert written.additionalMetadata == additional_metadata
+
+
+def testNonserializableDescriptionAndAdditionalMetadata(tmp_path):
+    output_file = tmp_path / 'test.db'
+    sink = large_image_source_zarr.new()
+    sink.addTile(np.zeros((256, 256, 1), dtype=np.uint8), x=0, y=0)
+
+    created = datetime.now(tz=timezone.utc)
+    with pytest.raises(TileSourceError):
+        sink.imageDescription = created
+    with pytest.raises(TileSourceError):
+        sink.additionalMetadata = dict(created=created)
+    sink.write(output_file)
+
+
+def testNoneDescriptionAndAdditionalMetadata():
+    sink = large_image_source_zarr.new()
+    assert sink.imageDescription is None
+    assert sink.additionalMetadata is None
+
+
+def testSingleBand(tmp_path):
+    output_file = tmp_path / 'test.db'
+    sink = large_image_source_zarr.new()
+    sink.addTile(np.zeros((1024, 1024, 1), dtype=np.uint8))
+    sink.write(output_file)
+
+
+def testSingleBandAndSingleX(tmp_path):
+    output_file = tmp_path / 'test.db'
+    sink = large_image_source_zarr.new()
+    sink.addTile(np.zeros((1, 1024, 1), dtype=np.uint8))
+    sink.write(output_file)
+
+
+@pytest.mark.parametrize(('n_gcps', 't_gcps'), [(2, list), (2, str), (4, list), (4, str)])
+def testGeoreferencing(tmp_path, n_gcps, t_gcps):
+    output_file = tmp_path / 'test.tiff'
+    sink = large_image_source_zarr.new()
+
+    sink.addTile(np.random.random((256, 256, 3)) * 10, x=0, y=0)
+
+    sink.projection = 4326
+    bbox = [
+        # covers lincoln memorial
+        [-77.050923, 38.888728],
+        [-77.049437, 38.889805],
+    ]
+    gcps = [
+        (bbox[0][0], bbox[0][1], 0, 0),
+        (bbox[0][0], bbox[1][1], 0, 256),
+        (bbox[1][0], bbox[1][1], 256, 256),
+        (bbox[1][0], bbox[0][1], 256, 0),
+    ]
+    if n_gcps == 2:
+        gcps = [gcps[0], gcps[2]]
+    if t_gcps is str:
+        gcps = ' '.join(' '.join(str(v) for v in gcp) for gcp in gcps)
+    sink.gcps = gcps
+    sink.write(output_file)
+
+    written = large_image.open(output_file)
+    metadata = written.getMetadata()
+    assert metadata['geospatial']
+    sb = metadata['sourceBounds']
+    assert sb['srs'] == '+proj=longlat +datum=WGS84 +no_defs'
+    assert sb['xmin'] == bbox[0][0]
+    assert sb['xmax'] == bbox[1][0]
+    assert sb['ymin'] == bbox[0][1]
+    assert sb['ymax'] == bbox[1][1]
+
+
+def testGeoreferencingNoProjection(tmp_path):
+    output_file = tmp_path / 'test.tiff'
+    sink = large_image_source_zarr.new()
+
+    sink.addTile(np.random.random((256, 256, 3)) * 10, x=0, y=0)
+
+    bbox = [
+        # covers lincoln memorial
+        [-77.050923, 38.888728],
+        [-77.049437, 38.889805],
+    ]
+    sink.gcps = [
+        (bbox[0][0], bbox[0][1], 0, 0),
+        (bbox[1][0], bbox[1][1], 256, 256),
+    ]
+    sink.write(output_file)
+    written = large_image.open(output_file)
+    metadata = written.getMetadata()
+    assert metadata.get('geospatial') is None
+
+
+def testGeoreferencingNoGCPs(tmp_path):
+    output_file = tmp_path / 'test.tiff'
+    sink = large_image_source_zarr.new()
+
+    sink.addTile(np.random.random((256, 256, 3)) * 10, x=0, y=0)
+
+    sink.projection = 4326
+    with pytest.raises(TileSourceError):
+        sink.write(output_file)
+    written = large_image.open(output_file)
+    metadata = written.getMetadata()
+    assert metadata.get('geospatial') is None
+
+
+@pytest.mark.parametrize('projection', ['', 'invalid', 'epsg:0', 0])
+def testGeoreferencingInvalidProjection(tmp_path, projection):
+    sink = large_image_source_zarr.new()
+
+    with pytest.raises(TileSourceError):
+        sink.projection = projection
+
+
+@pytest.mark.parametrize('gcps', [
+    '', 0, [], [()], [(), ()], ['50', '100'], [(-77, 38, 0, 0)], '-77 38 0 0',
+])
+def testGeoreferencingInvalidGCPs(tmp_path, gcps):
+    sink = large_image_source_zarr.new()
+
+    with pytest.raises(TileSourceError):
+        sink.gcps = gcps
+
+
+def testGeoreferencingDuplicateGCPs(tmp_path):
+    sink = large_image_source_zarr.new()
+    gcps = [
+        (-77, -38, 0, 0),
+        (-77, -38, 10, 10),
+        (-76, -39, 0, 0),
+        (-76, -39, 10, 10),
+    ]
+    with pytest.warns(Warning):
+        sink.gcps = gcps
+    assert sink.gcps == [(-77, -38, 0, 0), (-76, -39, 10, 10)]
